+++ src/utils/jobProfileNameCellFormatter.js
@@ -1,8 +1,15 @@
 import React from 'react';
 
-import { TextLink } from '@folio/stripes/components';
+import {
+  NoValue,
+  TextLink,
+} from '@folio/stripes/components';
 
 export const jobProfileNameCellFormatter = record => {
+  if (!record.jobProfileInfo) {
+    return <NoValue />;
+  }
+
   const { jobProfileInfo: { id, name } } = record;
   const linkToJobProfile = `/settings/data-import/job-profiles/view/${id}`;